requirements.txt
Flask==0.10.1
Jinja2==2.7.1
MarkupSafe==0.18
Pillow==2.2.2
Werkzeug==0.9.4
itsdangerous==0.23
nose==1.3.0
py==1.4.19
tinys3==0.1.9
tox==1.6.1
virtualenv==1.10.1
